Moves ExperimentHeader component to components/experiments folder
